@@ -1,41 +1,27 @@
 package com.trayis.simplikmvvm.ui
 
-import android.databinding.DataBindingUtil
-import android.databinding.ViewDataBinding
 import android.os.Bundle
-import android.support.v4.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.trayis.simplikmvvm.utils.Logging
+import androidx.databinding.DataBindingUtil
+import androidx.databinding.ViewDataBinding
+import androidx.fragment.app.Fragment
 import com.trayis.simplikmvvm.utils.SimpliProviderUtil
 import com.trayis.simplikmvvm.viewmodel.SimpliViewModel
-import java.util.*
 
-abstract class SimpliFragment<B : ViewDataBinding, V : SimpliViewModel> : Fragment(), Simpli {
+abstract class SimpliFragment<B : ViewDataBinding> : Fragment(), SimpliBase {
 
     protected val TAG = javaClass.simpleName
 
-    protected var mViewModel: V? = null
+    protected var mViewModels: Array<SimpliViewModel?>? = null
+
     protected var mBinding: B? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        mViewModel = getViewModel()
+        mViewModels = SimpliProviderUtil.instance.provider?.getViewModels(this)
         setHasOptionsMenu(false)
-    }
-
-    private fun getViewModel(): V? {
-        if (mViewModel == null) {
-            try {
-                @Suppress("UNCHECKED_CAST")
-                mViewModel = SimpliProviderUtil.instance.provider?.getViewModel(this) as V
-            } catch (e: InvalidPropertiesFormatException) {
-                Logging.e(TAG, e.message, e)
-            }
-
-        }
-        return mViewModel
     }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
@@ -46,12 +32,9 @@ abstract class SimpliFragment<B : ViewDataBinding, V : SimpliViewModel> : Fragme
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        bindViewModels()
         mBinding?.setLifecycleOwner(this)
-        val modelVariable = modelVariable
-        if (modelVariable > 0) {
-            mBinding?.setVariable(modelVariable, mViewModel)
-        }
-        mViewModel?.onCreate()
+        mViewModels?.forEach { it?.onCreate() }
         mBinding?.executePendingBindings()
     }
 
