@@ -1,12 +1,12 @@
 package com.trayis.simplikmvvm.utils
 
-import com.trayis.simplikmvvm.ui.Simpli
+import com.trayis.simplikmvvm.ui.SimpliBase
 import com.trayis.simplikmvvm.viewmodel.SimpliViewModel
 import java.util.*
 
-interface SimpliMvvmProvider<V : SimpliViewModel> {
+interface SimpliMvvmProvider {
 
     @Throws(InvalidPropertiesFormatException::class)
-    fun getViewModel(simpli: Simpli): V
+    fun getViewModels(base: SimpliBase): Array<SimpliViewModel?>?
 
 }
