@@ -2,7 +2,7 @@ package com.trayis.simplikmvvm.utils
 
 class SimpliProviderUtil private constructor() {
 
-    var provider: SimpliMvvmProvider<*>? = null
+    var provider: SimpliMvvmProvider? = null
         set(provider) {
             if (this.provider != null) {
                 Logging.w(TAG, "provider already set")
