@@ -1,10 +1,6 @@
 package com.trayis.simplimvvm.ui;
 
-import android.databinding.DataBindingUtil;
-import android.databinding.ViewDataBinding;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
 
 import com.trayis.simplimvvm.utils.Logging;
 import com.trayis.simplimvvm.utils.SimpliProviderUtil;
@@ -12,13 +8,30 @@ import com.trayis.simplimvvm.viewmodel.SimpliViewModel;
 
 import java.util.InvalidPropertiesFormatException;
 
-public abstract class SimpliActivity<B extends ViewDataBinding, V extends SimpliViewModel> extends AppCompatActivity implements Simpli {
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.databinding.DataBindingUtil;
+import androidx.databinding.ViewDataBinding;
+
+import static com.trayis.simplimvvm.ui.SimpliBase.ViewModelProviderFlag.SELF;
+
+public abstract class SimpliActivity<B extends ViewDataBinding> extends AppCompatActivity implements SimpliBase {
 
     protected final String TAG = getClass().getSimpleName();
 
     protected B mBinding;
 
-    protected V mViewModel;
+    protected SimpliViewModel[] mViewModels;
+
+    private int flag = SELF;
+
+    public int getViewModelProviderFlag() {
+        return flag;
+    }
+
+    protected void setViewModelProviderFlag(int flag) {
+        this.flag = flag;
+    }
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
@@ -28,26 +41,27 @@ public abstract class SimpliActivity<B extends ViewDataBinding, V extends Simpli
 
     private void doDataBinding() {
         mBinding = DataBindingUtil.setContentView(this, getLayoutResourceId());
-        mViewModel = getViewModel();
+        mViewModels = getViewModels();
+        setViewModelsToView();
         mBinding.setLifecycleOwner(this);
-        int modelVariable = getModelVariable();
-        if (modelVariable > 0) {
-            mBinding.setVariable(modelVariable, mViewModel);
+        if (mViewModels != null) {
+            for (SimpliViewModel viewModel : mViewModels) {
+                viewModel.onCreate();
+            }
         }
-        mViewModel.onCreate();
         mBinding.executePendingBindings();
     }
 
     @SuppressWarnings("unchecked")
-    private V getViewModel() {
-        if (mViewModel == null) {
+    private SimpliViewModel[] getViewModels() {
+        if (mViewModels == null) {
             try {
-                mViewModel = (V) SimpliProviderUtil.getInstance().getProvider().getViewModel(this);
+                mViewModels = SimpliProviderUtil.getInstance().getProvider().getViewModels(this);
             } catch (InvalidPropertiesFormatException e) {
                 Logging.e(TAG, e.getMessage(), e);
             }
         }
-        return mViewModel;
+        return mViewModels;
     }
 
 }
