@@ -1,11 +1,6 @@
 package com.trayis.simplimvvm.ui;
 
-import android.databinding.DataBindingUtil;
-import android.databinding.ViewDataBinding;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -16,12 +11,31 @@ import com.trayis.simplimvvm.viewmodel.SimpliViewModel;
 
 import java.util.InvalidPropertiesFormatException;
 
-public abstract class SimpliFragment<B extends ViewDataBinding, V extends SimpliViewModel> extends Fragment implements Simpli {
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.databinding.DataBindingUtil;
+import androidx.databinding.ViewDataBinding;
+import androidx.fragment.app.Fragment;
+
+import static com.trayis.simplimvvm.ui.SimpliBase.ViewModelProviderFlag.SELF;
+
+public abstract class SimpliFragment<B extends ViewDataBinding> extends Fragment implements SimpliBase {
 
     protected final String TAG = getClass().getSimpleName();
 
-    private V mViewModel;
-    private B mBinding;
+    private SimpliViewModel[] mViewModel;
+
+    protected B mBinding;
+
+    private int flag = SELF;
+
+    public int getViewModelProviderFlag() {
+        return flag;
+    }
+
+    protected void setViewModelProviderFlag(int flag) {
+        this.flag = flag;
+    }
 
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState) {
@@ -31,10 +45,10 @@ public abstract class SimpliFragment<B extends ViewDataBinding, V extends Simpli
     }
 
     @SuppressWarnings("unchecked")
-    private V getViewModel() {
+    private SimpliViewModel[] getViewModel() {
         if (mViewModel == null) {
             try {
-                mViewModel = (V) SimpliProviderUtil.getInstance().getProvider().getViewModel(this);
+                mViewModel = SimpliProviderUtil.getInstance().getProvider().getViewModels(this);
             } catch (InvalidPropertiesFormatException e) {
                 Logging.e(TAG, e.getMessage(), e);
             }
@@ -51,12 +65,13 @@ public abstract class SimpliFragment<B extends ViewDataBinding, V extends Simpli
     @Override
     public void onViewCreated(@NonNull View view, Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
+        setViewModelsToView();
         mBinding.setLifecycleOwner(this);
-        int modelVariable = getModelVariable();
-        if (modelVariable > 0) {
-            mBinding.setVariable(modelVariable, mViewModel);
+        if (mViewModel != null) {
+            for (SimpliViewModel viewModel : mViewModel) {
+                viewModel.onCreate();
+            }
         }
-        mViewModel.onCreate();
         mBinding.executePendingBindings();
     }
 
